refactor: update version check logic and refine `BaseSwitchFunctionHookItem`
* refactor: change `versionCode` parameter type from `Long` to `Int` in `HostInfo` and add debug logging
* fix: optimize `BaseSwitchFunctionHookItem` loading logic to prevent redundant operations and fix log tags
* feat: update `WeSettingInjector` to use `requireMinWeChatVersion` with `MMVersion` constant
* chore: remove obsolete VCS mappings in `.idea/vcs.xml`

+++ app/src/main/java/moe/ouom/wekit/host/HostInfo.java
@@ -66,7 +66,7 @@ public class HostInfo {
         return moe.ouom.wekit.host.impl.HostInfo.isWeChat();
     }
 
-    public static boolean requireMinWeChatVersion(long versionCode) {
+    public static boolean requireMinWeChatVersion(int versionCode) {
         return isWeChat() && getLongVersionCode() >= versionCode;
     }
 }
